\name{posterior_chisq_1}
\alias{posterior_chisq_1}
\title{Posterior Chi squared (method 1)}
\usage{
  posterior_chisq_1(y, yrep)
}
\arguments{
  \item{y}{Response variable}

  \item{yrep}{Draws from the posterior predictive
  distribution of \code{y}.}
}
\description{
  Posterior Chi squared (method 1)
}

