\name{posterior_chisq_2}
\alias{posterior_chisq_2}
\title{Posterior p-value for Chi Squared discrepancy (method 2)}
\usage{
  posterior_chisq_2(y, yrep, yhat)
}
\arguments{
  \item{y}{Data}

  \item{yrep}{Draws from the posterior predictive
  distribution of \code{y}.}

  \item{yhat}{Estimated value of \code{y}}
}
\description{
  Posterior p-value for Chi Squared discrepancy (method 2)
}

