\name{posterior_freeman_tukey}
\alias{posterior_freeman_tukey}
\title{Posterior p-value for Freeman-Tukey discrepancy}
\usage{
  posterior_freeman_tukey(y, yrep, yhat)
}
\arguments{
  \item{y}{Data}

  \item{yrep}{Draws from the posterior predictive
  distribution of \code{y}.}

  \item{yhat}{Estimated value of \code{y}}
}
\description{
  Posterior p-value for Freeman-Tukey discrepancy
}

