\name{predconcord}
\alias{predconcord}
\title{Predictive Concordance}
\usage{
  predconcord(y, yrep, qmin = 0.025, qmax = 0.975)
}
\arguments{
  \item{y}{Response variable}

  \item{yrep}{Draws from the posterior predictive
  distribution of \code{y}.}

  \item{qmin}{Minimum quantile of the range.}

  \item{qmax}{Maximum quantile of the range.}
}
\description{
  Predictive Concordance
}

