\name{InfoCriterion}
\alias{DIC}
\alias{DIC-class}
\alias{InfoCriterion}
\alias{InfoCriterion-class}
\alias{WAIC}
\alias{WAIC-class}
\title{Information Criterion Objects}
\description{
  Class for information criteria (AIC, DIC, WAIC)
}
\section{Slots}{
  \describe{ \item{\code{.Data}}{Value of the information
  criterion} \item{\code{loglik}}{Log likelihood.}
  \item{\code{b}}{bias correction.} \item{\code{n}}{number
  of observations.} }
}

